Add headers where needed to curl
I'm not sure this is a good change, but I had to stop and think about why the curl examples didn't work as written with my implementation, so maybe this lowers the cognitive load on the reader? We talk about the version header explicitly at the top, but if one is skimming, or coming in from google it's easy to miss... 

This change adds the broker api headers. These are required if one is using cloudfoundry-community service broker starter, and in general are good practice. One could write a broker that makes the examples work as they exist today, but since we're specific about what we accept I thought it would be nice to be explicit in what we send.