remove unnecessary conditionals
A couple of sections were enclosed in a conditional that checks if the
platform code is 'CF', i.e. if the context is Cloud Foundry. Within
those sections, there was a second conditional block (or a pair of them)
checking if the platform code was 'CF' or 'PCF'. No such check should be
necessary as both of these sections should be entirely absent in non-CF
documentation.